+++ android/src/main/kotlin/com/posthog/flutter/PostHogVersion.kt
@@ -1,3 +1,3 @@
 package com.posthog.flutter
 
-internal val postHogVersion = "5.11.1"
+internal val postHogVersion = "5.12.0"